Check if database exist

+++ flaskProject/server_create_db.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+set -e
 if [ -e "migrations" ]; then
     rm -r "migrations"
 fi
@@ -10,7 +11,11 @@ fi
 if [ -f "data.db" ]; then
     rm "data.db"
 fi
-psql -d $DATABASE_URL -c "DROP TABLE alembic_version ;"
+
+if [[ -n "$DATABASE_UR" ]]; then
+    psql -d $DATABASE_URL -c "DROP TABLE alembic_version ;"
+fi
+
 flask db init
 flask db migrate -m "Initial migration"
 flask db upgrade
